+++ tests/integration/models/reddit/test_subreddit.py
@@ -1046,6 +1046,35 @@ class TestSubredditStreams(IntegrationTest):
             for i in range(101):
                 assert isinstance(next(generator), Submission)
 
+    @mock.patch('time.sleep', return_value=None)
+    def test_pause_comments(self, _):
+        with self.recorder.use_cassette('TestSubredditStreams.'
+                                        'comments_pause'):
+            stream = self.reddit.subreddit('kakapo').stream
+            with_pause = stream.comments(pause_after=0)
+            was_paused = 0
+            for i in range(35):
+                item = next(with_pause)
+                assert isinstance(item, Comment) or item is None
+                if item is None:
+                    was_paused += 1
+            assert was_paused > 0
+
+    @mock.patch('time.sleep', return_value=None)
+    def test_resume_comments(self, _):
+        with self.recorder.use_cassette('TestSubredditStreams.'
+                                        'comments_resume'):
+            stream = self.reddit.subreddit('kakapo').stream
+            with_pause = stream.comments(pause_after=0)
+            for comment in with_pause:
+                if comment is None:
+                    break
+                assert isinstance(comment, Comment)
+            # The cassete was recorded with a period of no new comments,
+            # plus a new one at the end.
+            item = next(with_pause)
+            assert isinstance(item, Comment)
+
 
 class TestSubredditStylesheet(IntegrationTest):
     @staticmethod